Obtener de Jeo4j

+++ backend/models/carrera.go
@@ -1,9 +1,9 @@
 package models
 
 type Carrera struct {
-	Facultad               string  `json:"facultad"`
-	Nombre                 string  `json:"nombre_carrera"`
-	Director               string  `json:"director"`
-	Duracion               float64 `json:"duracion"`
-	EstudiantesRegistrados int     `json:"estudiantes_registrados"`
+	Facultad               string `json:"facultad"`
+	Nombre                 string `json:"nombre_carrera"`
+	Director               string `json:"director"`
+	Duracion               int64  `json:"duracion"`
+	EstudiantesRegistrados int64  `json:"estudiantes_registrados"`
 }
